requirements.txt
google-generativeai
requests
pyyaml
